+++ src/com/nixinova/graphics/Render3D.java
@@ -4,8 +4,8 @@ import java.util.Arrays;
 import java.util.stream.IntStream;
 
 import com.nixinova.coords.BlockCoord;
+import com.nixinova.coords.Coord;
 import com.nixinova.coords.PxCoord;
-import com.nixinova.coords.TxCoord;
 import com.nixinova.main.Game;
 import com.nixinova.player.Player;
 import com.nixinova.readwrite.Options;
@@ -30,7 +30,7 @@ public class Render3D extends Render {
 	public void renderWorld(Game game) {
 		final int TEX_SIZE = Conversion.PX_PER_BLOCK;
 
-		PxCoord pos = game.controls.getPositionPx();
+		PxCoord pos = game.controls.getPosition().toPx();
 		double bobbing = Math.sin(game.time) / 10.0;
 		double rotation = game.controls.getXRot();
 		double rotCos = Math.cos(rotation);
@@ -43,7 +43,7 @@ public class Render3D extends Render {
 		IntStream indicesRange = IntStream.range(0, this.lastKbdInput != null ? this.lastKbdInput.length : 0);
 		boolean hasntPressed = indicesRange.allMatch(i -> game.kbdInput[i] == this.lastKbdInput[i]); // check all array items are equal
 		boolean hasntMoved = pos.x == lastXMove && pos.y == lastYMove && pos.z == lastZMove;
-		boolean hasntFallen = lastGround == game.controls.getPlayerGroundPx();
+		boolean hasntFallen = lastGround == game.controls.getPlayerGround().toPx().y;
 		boolean hasntLooked = rotation == lastRot && tilt == lastTilt;
 		if (hasntPressed && hasntMoved && hasntFallen && hasntLooked) {
 			return;
@@ -69,12 +69,10 @@ public class Render3D extends Render {
 				double horiz = (xPx - this.width / 2.0D) / this.height;
 
 				// Depth calculation
-				double offset = pos.y;
+				double ground = game.controls.getPlayerGround().toSubBlock().y;
+				double offset = pos.y + ground;
 				double skyDepth = (World.SKY_Y_PX - offset) / -vert;
 				double worldDepth = (Player.PLAYER_HEIGHT + offset) / vert;
-				final int worldFallSpeed = 10;
-				double groundPx = game.controls.getPlayerGroundPx();
-				worldDepth += (groundPx - (int) groundPx) * worldFallSpeed; // world layer falling through effect
 				double depth = generateSky ? skyDepth : worldDepth;
 				if (game.controls.isWalking) {
 					depth += bobbing;
@@ -87,11 +85,11 @@ public class Render3D extends Render {
 
 				// World pixel coords
 				int texelX = (int) (horiz * rotCos + depth * rotSin + pos.x);
-				int texelY = game.controls.getPlayerGroundTx() * Conversion.PX_PER_BLOCK;
+				int texelY = (int) game.controls.getPlayerGround().toTx().y;
 				int texelZ = (int) (depth * rotCos - horiz * rotSin + pos.z);
 
 				// World block coords
-				BlockCoord blockCoord = new TxCoord(texelX, texelY, texelZ).toBlockCoord();
+				BlockCoord blockCoord = Coord.fromTx(texelX, texelY, texelZ).toBlock();
 				int blockX = blockCoord.x;
 				int blockY = blockCoord.y;
 				int blockZ = blockCoord.z;
@@ -124,7 +122,7 @@ public class Render3D extends Render {
 		this.lastXMove = pos.x;
 		this.lastYMove = pos.y;
 		this.lastZMove = pos.z;
-		this.lastGround = game.controls.getPlayerGroundPx();
+		this.lastGround = game.controls.getPlayerGround().toPx().y;
 		this.lastRot = rotation;
 		this.lastTilt = tilt;
 		this.lastKbdInput = Arrays.copyOf(game.kbdInput, game.kbdInput.length);